+++ tests/products/test_estimate_sm.py
@@ -6,7 +6,7 @@ from neptoon.products.estimate_sm import (
 )
 from neptoon.columns import ColumnInfo
 from neptoon.corrections.theory.neutrons_to_soil_moisture import (
-    neutrons_to_total_grav_soil_moisture_koehli_etal_2021,
+    neutrons_to_grav_soil_moisture_koehli_etal_2021,
 )
 from pandera.errors import SchemaError
 
@@ -185,10 +185,10 @@ def test_koehli_method_nans_not_processed():
     """
     Tests if the koehli method will process a nan value
     """
-    nan_neut = neutrons_to_total_grav_soil_moisture_koehli_etal_2021(
+    nan_neut = neutrons_to_grav_soil_moisture_koehli_etal_2021(
         neutron_count=np.nan, n0=2000, abs_air_humidity=8
     )
-    nan_hum = neutrons_to_total_grav_soil_moisture_koehli_etal_2021(
+    nan_hum = neutrons_to_grav_soil_moisture_koehli_etal_2021(
         neutron_count=1000, n0=2000, abs_air_humidity=np.nan
     )